+++ my-admin/src/api/api.ts
@@ -1,4 +1,6 @@
-export const getToken = async () => {
+import { fetchUtils } from "react-admin";
+
+const getToken = async () => {
   const token = await fetch("http://3.65.149.62/test-api/auth/login", {
     method: "POST",
     headers: {
@@ -10,4 +12,16 @@ export const getToken = async () => {
     }),
   });
   return await token.json();
+};
+
+export const httpClient = async (
+  url: string,
+  options: fetchUtils.Options = {}
+) => {
+  const token = await getToken().then((value) => value.accessToken);
+  const user = {
+    token: `Bearer ${token}`,
+    authenticated: !!token,
+  };
+  return fetchUtils.fetchJson(url, { ...options, user });
 };
