+++ my-admin/src/components/CustomPagination/CustomPagination.tsx
@@ -1,22 +1,26 @@
 import {
   Pagination,
-  PaginationActions as RaPaginationActions,
+  PaginationActions,
   PaginationActionsProps,
   PaginationProps,
+  useListContext,
 } from "react-admin";
 
-const PaginationActions = (props: PaginationActionsProps) => (
-  <RaPaginationActions {...props} />
-);
+const Actions = (props: PaginationActionsProps) => {
+  return <PaginationActions {...props} />;
+};
 
 const CustomPagination = (props: PaginationProps) => {
-  return (
-    <Pagination
-      ActionsComponent={PaginationActions}
-      rowsPerPageOptions={[]}
-      {...props}
-    />
-  );
+  const { page, setPage } = useListContext();
+
+  if (page >= 7) {
+    setPage(6);
+    const popUp = document.querySelector<HTMLElement>(".popup");
+    popUp!.style.opacity = "1";
+    popUp!.style.visibility = "visible";
+  }
+
+  return <Pagination rowsPerPageOptions={[]} actions={Actions} {...props} />;
 };
 
 export default CustomPagination;
